requirements.txt
# pip requirements file
# usage: pip install -r requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-WTF==0.14.2
PyMySQL==0.8.0
PyYAML==3.12
gunicorn==19.7.1
PyJWT==1.5.3